requirements.txt
about-time==3.1.1
alive-progress==2.4.1
autopep8==1.6.0
gitdb==4.0.9
GitPython==3.1.27
grapheme==0.6.0
pycodestyle==2.8.0
smmap==5.0.0
toml==0.10.2
